LIU-455: Move simple.py from DALiuGE apps to PyfuncApps
